1.0
Not work as i expected, in this code I've tried to count the occurence of the words when i read the file, so in that case I read the struct only in the Master, when i've all result, but in recive the master read only ~800, and i'dont understad why beacause before the recive, in the slave i've checked if it's all ok, and it is! So i don't know wyh when i recive the struct in the master why it read only a part of that.